+++ src/configuration/models/unite/uniteBuildConfiguration.ts
@@ -5,5 +5,4 @@ export class UniteBuildConfiguration {
     public bundle: boolean;
     public minify: boolean;
     public sourcemaps: boolean;
-    public variables: { [id: string]: any};
 }